Spinal case wip
Early code for the spinal casing algo.  The first regexp only selected spaces and did work for doing that but the actual answer requires you to replace any instances of uppercase characters with a "hyphen" by way of a space replacement, as well.  There's a commented regexp that i don't think quite does what i need it to do.